requirements.txt
fastapi==0.115.4
uvicorn==0.32.0
psycopg[binary]==3.2.3
python-dotenv==1.0.1
